+++ src/modules/chorus/interfaces/cli/chorus.cli.controller.ts
@@ -1,10 +1,10 @@
 import fs from "fs";
-import path from "path";
 
 import { StaticImplements } from "../../../../decorators/staticImplements.decorator";
 import { CliStaticInterface} from "../../../../@types/Cli.interface";
 import ChorusParser from "../../chorus.parser";
 import chorusService, { RejectedRequest } from "../../chorus.service";
+import { findFiles } from "../../../../shared/helpers/ParserHelper";
 
 @StaticImplements<CliStaticInterface>()
 export default class ChorusCliController {
@@ -24,16 +24,7 @@ export default class ChorusCliController {
             throw new Error(`File not found ${file}`);
         }
 
-        const files = [];
-
-        if (fs.lstatSync(file).isDirectory()) {
-            const filesInFolder = fs
-                .readdirSync(file)
-                .filter(fileName => !fileName.startsWith(".") && !fs.lstatSync(path.join(file, fileName)).isDirectory())
-                .map((fileName => path.join(file, fileName)));
-
-            files.push(...filesInFolder);
-        } else files.push(file);
+        const files = findFiles(file);
 
         console.info(`${files.length} files in the parse queue`);
         console.info(`You can read log in ${this.logFileParsePath}`);
