+++ src/shared/ComptesAccepted.ts
@@ -1,0 +1,4 @@
+export const COMPTES_ACCEPTED = [
+    "TD ASSOCIATION", 
+    "TI EP ASSO GIP"
+]
