+++ src/shared/helpers/SirenHelper.ts
@@ -1,0 +1,5 @@
+import { Siret } from "../../@types/Siret";
+
+export function siretToSiren(siret: Siret) {
+    return siret.slice(0,9);
+}
